+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### TOFcam670
+- Added interference filter
 - Added edge filter
 
 ## [0.9.0] - 2026-05-20